+++ src/main/java/wooteco/subway/ui/SectionController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import wooteco.subway.service.SectionService;
-import wooteco.subway.ui.dto.SectionRequest;
+import wooteco.subway.ui.dto.request.SectionRequest;
 
 @RestController
 @RequestMapping("/lines/{lineId}/sections")
