+++ src/test/java/wooteco/subway/acceptance/StationAcceptanceTest.java
@@ -12,9 +12,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.jdbc.Sql;
-import wooteco.subway.ui.dto.ExceptionResponse;
-import wooteco.subway.ui.dto.StationRequest;
-import wooteco.subway.ui.dto.StationResponse;
+import wooteco.subway.ui.dto.request.StationRequest;
+import wooteco.subway.ui.dto.response.ExceptionResponse;
+import wooteco.subway.ui.dto.response.StationResponse;
 
 @DisplayName("지하철역 E2E")
 @Sql("classpath:/schema.sql")
